Update how GitHub Issue Templates are updated
Update how the `process-toolkit-schema.ts` automation tool
generates and updates GitHub Issue Templates, ensuring that the
modified file maintains the existing organization and formatting
of the original.

Additionally, this change fixes a bug that caused the
contents of the "Toolkit Tool Issue Report" Issue Template
to be written to the "Feature Request or Suggestion"
Issue Template instead.