Clean up and update automation tools
Clean up and update the `utils.ts` module and
the `process-toolkit-schema.ts` and `add-namespace.ts`
automation tools, including:
- Add missing documentation comments.
- Combine the functionality of `commitChanges()` and `printChanges()`
into `commitChanges()` in `process-toolkit-schema.ts`.
- Add the `tempDirPrefix` option and the ability to pass a
`WriteFileOptions` object to `writeFile()`.
- Update the behavior of temporary files in `writeFile()`,
including re-using the same temporary directories for all files
being updated by the script, storing all temporary files within a
subdirectory of the `typescript-toolkit` directory within the
OS's temporary file directory, and removing all of the aforementioned
subdirectories on program shutdown.
- Add the `tempDirPrefix` argument to `updateToolkitSchema()`.
- Update `process-toolkit-schema.ts` and `add-namespace.ts` to
match the documentation in the `README.md` file, as well as the
changes made to the `ToolkitSchema` in 9bc1e9f.
- Fix order of options for `add-namespace` in the `README.md`.